+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.3.34] - 2026-02-03
+
 ### Added
 
 - `pcb preview <path/to/board.zen>` to generate a preview link for a release.
@@ -264,7 +266,8 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 - Error on invalid type passed to `io()`
 - Format the auto-generated component .zen files
 
-[Unreleased]: https://github.com/diodeinc/pcb/compare/v0.3.33...HEAD
+[Unreleased]: https://github.com/diodeinc/pcb/compare/v0.3.34...HEAD
+[0.3.34]: https://github.com/diodeinc/pcb/compare/v0.3.33...v0.3.34
 [0.3.33]: https://github.com/diodeinc/pcb/compare/v0.3.32...v0.3.33
 [0.3.32]: https://github.com/diodeinc/pcb/compare/v0.3.31...v0.3.32
 [0.3.31]: https://github.com/diodeinc/pcb/compare/v0.3.30...v0.3.31